Add index error for goto card out of bounds
*Summary*
Prior to this change, using GotoCard with an out of bounds index
resulted in a python backtrace, hard-to-read (for new programmers)
error, and the error *not* being added to the errors buffer (because it
happened on main thread). After the change, error message is descriptive
and added to errors buffer.

*Test Plan*
To reproduce, simply use GotoCard(100) in an event handler in the
default/readme stack